stop process on interrupt
